model_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=2.2.0
transformers>=4.42.0
numpy>=1.26.0
python-dotenv==1.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
pydantic>=2.0.0
